+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- 2 tests for Tokenizer.tokenbuf
+- Tokenizer.tokenbuf push/consume methods
 - 2 tests for fmt::Debug on Token and TokenBody
 - fmt::Debug implementation for Token and TokenBody
 - Token, TokenBody, Span structs